rename connectors overview folder
