+++ src/components/ConnectorsOverview/DeleteConnectorModal/DeleteIntegrationModal.tsx
