+++ src/components/ConnectorsOverview/DeleteConnectorModal/index.ts
